refactor: replace notifications job with background worker
